Add auth to paths
TODO: need to find a way to trigger server action (Or other fetch) to be
able to reset cookies in case of refresh. Calling functions in server
component is not, in fact, a server action. Most obvious way to handle
this is in middleware, but that runs in the edge runtime (i.e. not in a
runtime that handles bcrypt, pg, crypto, &c.). Seems obvious way to go
about it is making calls to an endpoint in the middleware?